// src/modules/mgmt-api-client/model/contractAgreementDto.ts
/**
 * EDC REST API
 * EDC REST APIs - merged by OpenApiMerger
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Policy } from './policy';


export interface ContractAgreementDto { 
    assetId: string;
    consumerAgentId: string;
    contractEndDate?: number;
    contractSigningDate?: number;
    contractStartDate?: number;
    id: string;
    policy: Policy;
    providerAgentId: string;
}