// src/modules/mgmt-api-client/model/transferRequestDto.ts
/**
 * EDC REST API
 * EDC REST APIs - merged by OpenApiMerger
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DataAddress } from './dataAddress';
import { TransferType } from './transferType';


export interface TransferRequestDto { 
    assetId: string;
    connectorAddress: string;
    connectorId: string;
    contractId: string;
    dataDestination: DataAddress;
    id?: string;
    managedResources?: boolean;
    properties?: { [key: string]: string; };
    protocol: string;
    transferType: TransferType;
}